Update cfp tests to use size_t

+++ tests/cfp/testCfpArray1_source.c
@@ -75,7 +75,7 @@ _catFunc3(given_, CFP_ARRAY_TYPE, _when_get_expect_entryReturned)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_NAMESPACE.SUB_NAMESPACE.set(cfpArr, i, VAL);
 
   // dirty cache doesn't immediately apply compression
@@ -87,7 +87,7 @@ _catFunc3(given_, CFP_ARRAY_TYPE, _when_ref_expect_arrayObjectValid)(void **stat
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.ref(cfpArr, i);
 
   assert_ptr_equal(cfpArrRef.array.object, cfpArr.object);
@@ -98,7 +98,7 @@ _catFunc3(given_, CFP_ARRAY_TYPE, _when_ptr_expect_arrayObjectValid)(void **stat
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
 
   assert_ptr_equal(cfpArrPtr.reference.array.object, cfpArr.object);
@@ -136,7 +136,7 @@ _catFunc3(given_, CFP_REF_TYPE, _when_get_expect_entryReturned)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.ref(cfpArr, i);
   CFP_NAMESPACE.SUB_NAMESPACE.set(cfpArr, i, VAL);
 
@@ -148,7 +148,7 @@ _catFunc3(given_, CFP_REF_TYPE, _when_set_expect_arrayUpdated)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.ref(cfpArr, i);
   CFP_NAMESPACE.SUB_NAMESPACE.reference.set(cfpArrRef, VAL);
 
@@ -160,7 +160,7 @@ _catFunc3(given_, CFP_REF_TYPE, _when_copy_expect_arrayUpdated)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 2;
+  size_t i1 = 1, i2 = 2;
   CFP_NAMESPACE.SUB_NAMESPACE.set(cfpArr, i1, VAL);
   CFP_REF_TYPE cfpArrRef_a = CFP_NAMESPACE.SUB_NAMESPACE.ref(cfpArr, i1);
   CFP_REF_TYPE cfpArrRef_b = CFP_NAMESPACE.SUB_NAMESPACE.ref(cfpArr, i2);
@@ -174,7 +174,7 @@ _catFunc3(given_, CFP_REF_TYPE, _when_ptr_expect_addressMatches)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.ref(cfpArr, i);
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.reference.ptr(cfpArrRef);
 
@@ -191,7 +191,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_get_set_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   SCALAR val = 5;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   CFP_NAMESPACE.SUB_NAMESPACE.pointer.set(cfpArrPtr, val);
@@ -205,7 +205,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_get_at_set_at_expect_correct)(void **state
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1, io = 3;
+  size_t i = 1, io = 3;
   SCALAR val = 5;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   CFP_NAMESPACE.SUB_NAMESPACE.pointer.set_at(cfpArrPtr, io, val);
@@ -219,7 +219,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_ref_expect_addressMatches)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.pointer.ref(cfpArrPtr);
 
@@ -231,8 +231,8 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_ref_at_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
-  uint oi = 10;
+  size_t i = 1;
+  size_t oi = 10;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.pointer.ref_at(cfpArrPtr, oi);
 
@@ -245,7 +245,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_lt_expect_less)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 5;
+  size_t i1 = 1, i2 = 5;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
   CFP_PTR_TYPE cfpArrPtrB = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i2);
 
@@ -257,7 +257,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_gt_expect_greater)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 5;
+  size_t i1 = 1, i2 = 5;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
   CFP_PTR_TYPE cfpArrPtrB = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i2);
 
@@ -269,7 +269,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_leq_expect_less_or_equal)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 5;
+  size_t i1 = 1, i2 = 5;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
   CFP_PTR_TYPE cfpArrPtrB = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i2);
 
@@ -282,7 +282,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_geq_expect_greater_or_equal)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 5;
+  size_t i1 = 1, i2 = 5;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
   CFP_PTR_TYPE cfpArrPtrB = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i2);
 
@@ -295,7 +295,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_eq_expect_equal)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1;
+  size_t i1 = 1;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
 
   assert_true(CFP_NAMESPACE.SUB_NAMESPACE.pointer.eq(cfpArrPtrA, cfpArrPtrA));
@@ -306,7 +306,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_neq_expect_not_equal)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 5;
+  size_t i1 = 1, i2 = 5;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
   CFP_PTR_TYPE cfpArrPtrB = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i2);
 
@@ -318,7 +318,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_distance_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i1 = 1, i2 = 5;
+  size_t i1 = 1, i2 = 5;
   CFP_PTR_TYPE cfpArrPtrA = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i1);
   CFP_PTR_TYPE cfpArrPtrB = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i2);
 
@@ -331,7 +331,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_next_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1, oi = 10;
+  size_t i = 1, oi = 10;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.pointer.next(cfpArrPtr, oi);
 
@@ -344,7 +344,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_prev_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 15, oi = 10;
+  size_t i = 15, oi = 10;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.pointer.prev(cfpArrPtr, oi);
 
@@ -357,7 +357,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_inc_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.pointer.inc(cfpArrPtr);
 
@@ -370,7 +370,7 @@ _catFunc3(given_, CFP_PTR_TYPE, _when_dec_expect_correct)(void **state)
 {
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
-  uint i = 1;
+  size_t i = 1;
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.ptr(cfpArr, i);
   cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.pointer.dec(cfpArrPtr);
 
@@ -402,7 +402,7 @@ _catFunc3(given_, CFP_ITER_TYPE, _when_ref_at_expect_correct)(void **state)
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
 
-  uint io = 5;
+  size_t io = 5;
 
   CFP_ITER_TYPE cfpArrIter = CFP_NAMESPACE.SUB_NAMESPACE.begin(cfpArr);
   CFP_REF_TYPE cfpArrRef = CFP_NAMESPACE.SUB_NAMESPACE.iterator.ref_at(cfpArrIter, io);
@@ -430,7 +430,7 @@ _catFunc3(given_, CFP_ITER_TYPE, _when_ptr_at_expect_correct)(void **state)
   struct setupVars *bundle = *state;
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
 
-  uint io = 5;
+  size_t io = 5;
 
   CFP_ITER_TYPE cfpArrIter = CFP_NAMESPACE.SUB_NAMESPACE.begin(cfpArr);
   CFP_PTR_TYPE cfpArrPtr = CFP_NAMESPACE.SUB_NAMESPACE.iterator.ptr_at(cfpArrIter, io);
@@ -574,7 +574,7 @@ _catFunc3(given_, CFP_ITER_TYPE, _when_get_index_expect_correct)(void **state)
   CFP_ARRAY_TYPE cfpArr = bundle->cfpArr;
 
   CFP_ITER_TYPE cfpArrIter = CFP_NAMESPACE.SUB_NAMESPACE.begin(cfpArr);
-  int idx = CFP_NAMESPACE.SUB_NAMESPACE.iterator.i(cfpArrIter);
+  size_t idx = CFP_NAMESPACE.SUB_NAMESPACE.iterator.i(cfpArrIter);
 
-  assert_int_equal(idx, 0);
+  assert_int_equal(idx, 0u);
 }
